+++ src/components/control/index.js
@@ -1,7 +1,53 @@
 import React from 'react';
+import { useSelector, useDispatch } from "react-redux";
 
-const Control = props => {
-    const {  wantedListToShow, hideShowAllItems, itemsToShow, toggleItems, deleteCompletedItem, clearList, allCompletedItem } = props;
+import { 
+  setList, 
+  setWantedListToShow, 
+  setToggleItems, 
+} from "./../../actions/actionCreater.js";
+
+const Control = () => {
+
+    const { list, wantedListToShow, toggleItems } = useSelector(state => state);
+    const dispatch = useDispatch();
+
+    function itemsToShow(choise){
+        if(choise === 'active') {
+          dispatch(setWantedListToShow(list.filter(item => !item.complete)));
+        }else if(choise === 'completed'){
+          dispatch(setWantedListToShow(list.filter(item => item.complete)));
+        }else if(choise === 'all'){
+          dispatch(setWantedListToShow(list));
+        }
+     }
+    
+    function deleteCompletedItem(){
+        let newList = list.filter(item => !item.complete);
+        dispatch(setList(newList));
+        dispatch(setWantedListToShow(newList));
+    }
+
+    function clearList(){
+        dispatch(setList([]));
+        dispatch(setWantedListToShow([]));
+    }
+
+    function hideShowAllItems(){
+        toggleItems ? dispatch(setWantedListToShow(list)) : dispatch(setWantedListToShow([]));
+        dispatch(setToggleItems(!toggleItems));
+    }
+
+    function allCompletedItem(){
+        const listCopy = [...list];
+        let newList = listCopy.map(val => {
+            val.complete = true;
+            return val;
+        });
+        dispatch(setList(newList));
+        dispatch(setWantedListToShow(newList));
+    }
+
     return (
         <div className='text-center'>
             <button className='btn btn-primary mb-3 mr-3' onClick={hideShowAllItems}>
